+++ test_case/test_projection.h
@@ -89,8 +89,8 @@ void test_PlaneProjector(){
   PlaneProjector<double>::convert(planes, planes_for_each_node, x.size()/3);
   assert(findFeasible(planes_for_each_node, x));
   PlaneProjector<double> P(planes_for_each_node, x);
-  assert_eq(P.getFace().size(),x.size());
-  for (int i = 0; i < P.getFace().size(); ++i)
+  assert_eq((int)P.getFace().size(),(int)x.size());
+  for (int i = 0; i < (int)P.getFace().size(); ++i)
     assert_eq(P.getFace()[i],0);
 
   // test step limit
@@ -235,8 +235,8 @@ void test_OnePlaneProjector(){
   PlaneProjector<double>::convert(planes, planes_for_each_node, x.size()/3);
   assert(findFeasible(planes_for_each_node, x));
   PlaneProjector<double> P(planes_for_each_node, x);
-  assert_eq(P.getFace().size(),x.size());
-  for (int i = 0; i < P.getFace().size(); ++i)
+  assert_eq((int)P.getFace().size(),(int)x.size());
+  for (int i = 0; i < (int)P.getFace().size(); ++i)
     assert_eq(P.getFace()[i],0);
 
   // test step limit
